indexing_service/requirements.txt
# PyTorch with CUDA support
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.5.1

# Core dependencies
transformers==4.36.2
numpy==1.24.3
huggingface-hub==0.19.4
protobuf==5.29.2
sentencepiece==0.1.99  # Required for tokenizers

# Database
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0  # Required for async PostgreSQL support

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1

# Testing
pytest-asyncio==0.25.1 